Change cluster time type to double

+++ Detectors/Upgrades/ALICE3/IOTOF/DataFormatsIOTOF/include/DataFormatsIOTOF/Cluster.h
@@ -27,7 +27,7 @@ struct Cluster {
   int16_t subDetID = -1;
   int16_t layer = -1;
   int16_t disk = -1;
-  float time = 0.f;
+  double time = 0.0;
 
   std::string asString() const;
